refactor(ui): clean up sidebar — remove All, add Calendar + Media to Content, Settings under Intelligence
- Remove "All" from Content section (redundant with sub-pages)
- Calendar and Media moved into Content section
- Settings section directly below Intelligence (not at bottom)
- Monitor + CMO Chat stay at bottom as quick-access tools
- Extract ContentLibrary as shared component
- /content redirects to /content/articles
- "+ New" button on each content type page
